Backend/requirements.txt
fastapi
uvicorn[standard]
google-generativeai
qdrant-client
python-dotenv